refactor: Move ChatGPT interceptor to separate module and remove console logs
- Extract setupChatGPTInterceptor to lib/setupChatGPTInterceptor.ts
- Remove all console.log statements from background.ts
- Clean up imports and code organization
